Finished Exercises after chapter 2

+++ 02_Basics/MathCalculator/Exercise/lib.h
@@ -4,4 +4,4 @@
 
 double calculate_pi(const std::uint32_t num_iterations);
 
-void print_dec_to_bin(std::uint8_t value);
+void print_dec_to_bin(std::uint16_t value);
